Towards a new wags (#51)
This is the wags `1.0.0` branch. It is a complete rewrite of the `purescript-wags` project designed from the ground-up with an event-based architecture. While there is a lot to do to make this project as full-featured and well-tested as the older version of wags, it is well worth the redesign, as it is much faster and much more ergonomic.