Updates to lockless syntax (#66)
Removes the `lock` type parameter. This takes away the type-level guarantee that requires portals to be constrained to a computational context. Instead, at runtime, portals that have escaped their context will be recreated in the new context.

The motivation behind this decision is that the `lock` type parameter made working with `Element`s unergonomic, as the impredicative type would often cause unification issues for non-trivial function signatures. OTOH, folks rarely use portals, and when they do, they never run into the issue that locks were intended to address. So, even though this solution is a bit less conceptually elegant than the type-level one, it makes working with the framework more fluid.